+++ substrafl/algorithms/pytorch/torch_base_algo.py
@@ -2,6 +2,7 @@ import abc
 import inspect
 import logging
 import os
+import random
 import shutil
 from pathlib import Path
 from typing import Any
@@ -9,6 +10,7 @@ from typing import Optional
 from typing import Union
 
 import numpy as np
+import numpy.random
 import torch
 
 from substrafl.algorithms.algo import Algo
@@ -58,6 +60,10 @@ class TorchAlgo(Algo):
         super().__init__(*args, **kwargs)
 
         if seed is not None:
+            # For a better control of randomness, it is recommended to set python and NumPy seed along with PyTorch's:
+            # https://pytorch.org/docs/2.0/notes/randomness.html#controlling-sources-of-randomness
+            random.seed(seed)
+            np.random.seed(seed)
             torch.manual_seed(seed)
 
         self._device = self._get_torch_device(use_gpu=use_gpu)
@@ -273,10 +279,13 @@ class TorchAlgo(Algo):
 
         self._index_generator = checkpoint.pop("index_generator")
 
+        random.setstate(checkpoint.pop("random_rng_state"))
+        numpy.random.set_state(checkpoint.pop("numpy_rng_state"))
+
         if self._device == torch.device("cpu"):
-            torch.set_rng_state(checkpoint.pop("rng_state").to(self._device))
+            torch.set_rng_state(checkpoint.pop("torch_rng_state").to(self._device))
         else:
-            torch.cuda.set_rng_state(checkpoint.pop("rng_state").to("cpu"))
+            torch.cuda.set_rng_state(checkpoint.pop("torch_rng_state").to("cpu"))
 
         return checkpoint
 
@@ -324,10 +333,14 @@ class TorchAlgo(Algo):
         if self._scheduler is not None:
             checkpoint["scheduler_state_dict"] = self._scheduler.state_dict()
 
+        checkpoint["random_rng_state"] = random.getstate()
+
+        checkpoint["numpy_rng_state"] = np.random.get_state()
+
         if self._device == torch.device("cpu"):
-            checkpoint["rng_state"] = torch.get_rng_state()
+            checkpoint["torch_rng_state"] = torch.get_rng_state()
         else:
-            checkpoint["rng_state"] = torch.cuda.get_rng_state()
+            checkpoint["torch_rng_state"] = torch.cuda.get_rng_state()
 
         return checkpoint